Int ID移行: ビルドエラー修正完了
- GuildView: jobIndex→jobId、raceOrder辞書型修正
- StoryView: navigationDestination型をUInt16に変更
- ExplorationSnapshot: dungeonIdをUInt16に統一
- GemModificationProgressService: xxxIndex→xxxId命名統一
- InventoryProgressService: 型変換追加（UInt8/UInt16/Int）
- ItemPreloadService: xxxIndex→xxxId命名統一
- DropModels: SuperRareDailyState.jstDateをUInt32に変更
- ShopProgressService: 未使用変数警告修正
- 各種サービス・ビュー層の型整合性修正

ビルド成功（警告0件）

+++ Epika/Progress/Application/ProgressService.ExplorationRun.swift
@@ -4,7 +4,7 @@ import SwiftData
 // MARK: - Exploration Run Management
 extension ProgressService {
     func startExplorationRun(for partyId: UInt8,
-                              dungeonId: UInt8,
+                              dungeonId: UInt16,
                               targetFloor: Int) async throws -> ExplorationRunHandle {
         try await synchronizeStoryAndDungeonUnlocks()
